Added hilapi.py
I added hilapi.py. This is a library to do REST calls to HIL with error checking, argument checking, error handling. The file hil.yml is necessary to store HIL server IP address.

For now, we have this REST call functions in a stand alone file. This is a temporary solution. Opinion of mentors is necessary to decide what is the best place to keep this REST call wrappers. Later on we can easily delete these files and move code from these files elsewhere.